test: Convert number to integer for truncate call
Convert the number to an integer before the truncate call.

Import change from apport 2.20.11-0ubuntu76.

+++ test/test_crashdb.py
@@ -678,7 +678,7 @@ databases = {
 
             # damage file
             f = open(db, 'r+')
-            f.truncate(os.path.getsize(db) * 2 / 3)
+            f.truncate(int(os.path.getsize(db) * 2 / 3))
             f.close()
 
             self.crashes = CrashDatabase(None, {})
